requirements.txt
streamlit
pandas
numpy
matplotlib
matplotlib-ternary
plotly
scikit-learn
openpyxl
Pillow
streamlit-lottie
